+++ plugins/jnario-maven-plugin/src/main/java/org/jnario/maven/SuiteMavenModule.java
@@ -1,0 +1,13 @@
+package org.jnario.maven;
+
+import org.eclipse.xtend.maven.MavenTraceURIConverter;
+import org.eclipse.xtext.generator.trace.ITraceURIConverter;
+import org.jnario.suite.SuiteRuntimeModule;
+
+public class SuiteMavenModule extends SuiteRuntimeModule{
+
+	public Class<? extends ITraceURIConverter> bindITraceURIConverter() {
+		return MavenTraceURIConverter.class;
+	}
+	
+}